Remove duplicate dependencies favoring first fetched
Duplicate dependencies are removed after fetching all of them.
It would be better if they
weren't fetched in the first place.

Because of the function used to remove duplicates, the first fetched dependencies are favored
over dependencies fetched later (i.e. deps appearing earlier in the list are
prioritized over those appearing later)